+++ test/fixture/define/dynamic-import/output.js
@@ -1,5 +1,7 @@
-export default function fetch () {
-  return import('./foo')
-      .then(() => { console.log('success') })
-      .catch(() => { console.log('error') })
+export default function fetch() {
+  return import("./foo").then(() => {
+    console.log("success");
+  }).catch(() => {
+    console.log("error");
+  });
 }