+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # @buxlabs/amd-to-es6 changelog
 
 ## 0.13.0
+* add: support dynamic imports inside of the source code
 * edit: update dependencies
 * remove: comments support